Conflicts:
	configure
	mk/crates.mk
	mk/main.mk
	mk/platform.mk
	mk/prepare.mk
	mk/rt.mk
	mk/target.mk
	src/compiletest/header.rs
	src/libcore/intrinsics.rs
	src/librustc_back/archive.rs
	src/librustc_back/lib.rs
	src/librustc_back/target/mod.rs
	src/librustc_driver/driver.rs
	src/librustc_llvm/lib.rs
	src/librustc_trans/back/link.rs
	src/librustc_trans/trans/context.rs
	src/librustc_trans/trans/debuginfo/metadata.rs
	src/libstd/dynamic_lib.rs
	src/libstd/rt/libunwind.rs
	src/libstd/rt/unwind.rs
	src/libstd/sys/unix/thread.rs
	src/libstd/sys/unix/time.rs
	src/libstd/thread/local.rs
	src/libtest/lib.rs
	src/rt/rust_builtin.c
	src/rustllvm/ExecutionEngineWrapper.cpp
	src/test/run-pass-fulldeps/issue-15149.rs
	src/test/run-pass/out-of-stack-new-thread-no-split.rs

+++ src/test/run-pass-fulldeps/issue-15149.rs
@@ -9,7 +9,7 @@
 // except according to those terms.
 
 // no-prefer-dynamic
-// ignore-android
+// ignore-cross-compile
 // ignore-nacl
 
 #![feature(rustc_private)]
@@ -65,6 +65,10 @@ fn test() {
                     str::from_utf8(&child_output.stdout).unwrap(),
                     str::from_utf8(&child_output.stderr).unwrap()));
 
-    fs::remove_dir_all(&child_dir).unwrap();
-
+    let res = fs::remove_dir_all(&child_dir);
+    if res.is_err() {
+        // On Windows deleting just executed mytest.exe can fail because it's still locked
+        std::thread::sleep_ms(1000);
+        fs::remove_dir_all(&child_dir).unwrap();
+    }
 }
